CV is faster also for the laplacian
	modified:   ../tests/speedTests/ImageProcessing.h

git-svn-id: https://svnknoll.informatik.tu-muenchen.de/kloses/visionTools@916 f7bcf5df-4c73-0410-a0d8-b61c104701fe

+++ tests/speedTests/ImageProcessing.h
@@ -88,7 +88,7 @@ void testConvolutionOCV()
 
 	t.reset();
 	for( size_t i = 0; i <  NUMSAMPLES; i++ ){
-		cv::Laplacian( img, out, CV_32F );
+		cv::Laplacian( img, out, CV_8U );
 	}
 	std::cout << "OCV:\tLaplace_3x3\t-> avg. " << t.elapsedMilliSeconds() / NUMSAMPLES << "ms" << std::endl;
